ai-agent/requirements.txt
flask
crewai
crewai-tools
python-dotenv
google-generativeai
litellm